Add files via upload
First two games. One it's so easy to do and so easy to complete, the second one it's a little more complicated coded because of the use of sentences like while and te try to make a "do-while", but it's very great and it´s funny (at least for me). :)